Move principal-related fields into subdir
