+++ indico/web/client/js/react/components/index.js
@@ -29,4 +29,4 @@ export {default as Carousel} from './Carousel';
 export {default as PrincipalSearchField} from './PrincipalSearchField';
 export {default as ScrollButton} from './ScrollButton';
 export {default as StickyWithScrollBack} from './StickyWithScrollBack';
-export {default as PrincipalListField} from './PrincipalListField';
+export {PrincipalListField} from './principals';
