+++ indico/web/client/js/react/components/principals/GroupSearch.jsx
