+++ indico/web/client/js/react/components/principals/PrincipalListField.jsx
