+++ indico/web/client/js/react/components/principals/UserSearch.jsx
